// frontend-angular/src/app/api/models/ticket-create.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TicketState } from '../models/ticket-state';
import { TicketType } from '../models/ticket-type';
export interface TicketCreate {
  body: string;

  /**
   * Must belong to teamId (§6)
   */
  epicId?: number | null;

  /**
   * Optional; defaults to `new`
   */
  state?: TicketState;
  teamId: number;
  title: string;
  type: TicketType;
}